+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/Rule.java
@@ -23,7 +23,7 @@ public interface Rule {
     String getRuleName();
 
     @NonNull
-    String getRuleCategory();
+    String getRuleModuleId();
 
     @ServiceFilter
     boolean isRuleAvailable();