+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/Versioning.java
@@ -15,5 +15,7 @@ public interface Versioning {
 
     @NonNull String getVersioningName();
 
+    @NonNull String getVersioningModuleId();
+
     boolean isValidVersion(@NonNull CharSequence text);
 }